requirements.txt
pandas
numpy
faker
psycopg2-binary
tqdm
python-dateutil
